requirements.txt
asgiref==3.2.10
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
Markdown==3.2.2
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
